+++ package.nix
@@ -12,7 +12,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "forgemux";
-  version = "0.1.11";
+  version = "0.1.12";
 
   src = ./.;